Rename "version.py" to "constants.py".
