+++ pymcbdsc/constants.py
